+++ tests/test_packages/test_cleveref.py
@@ -81,7 +81,7 @@ plain_4 = """
 A LATEXXXERROR B
 """
 
-stderr_4 = r"""*** LaTeX error: line 3, column 2:
+stderr_4 = r"""*** LaTeX error: code in 't.tex', line 3, column 2:
 *** To use cleveref with YaLafi, you should use
 *** \YYCleverefInput to load the sed file, e.g.
 *** '\YYCleverefInput{main.sed}' if your LaTeX
@@ -93,7 +93,7 @@ def test_4(capsys):
     capsys.readouterr()
     parms = parameters.Parameters()
     p = parser.Parser(parms)
-    toks = p.parse(latex_4)
+    toks = p.parse(latex_4, source='t.tex')
     plain, pos = utils.get_txt_pos(toks)
     captured = capsys.readouterr()
     assert plain_4 == plain
@@ -112,7 +112,7 @@ plain_5 = """
 A LATEXXXERROR B
 """
 
-stderr_5 = r"""*** LaTeX error: line 4, column 2:
+stderr_5 = r"""*** LaTeX error: code in 't.tex', line 4, column 2:
 *** No replacement for \cref{lab} known.
 *** Run LaTeX again to build a new sed file.
 
@@ -124,7 +124,7 @@ def test_5(capsys):
     capsys.readouterr()
     parms = parameters.Parameters()
     p = parser.Parser(parms, read_macros=read)
-    toks = p.parse(latex_5)
+    toks = p.parse(latex_5, source='t.tex')
     plain, pos = utils.get_txt_pos(toks)
     captured = capsys.readouterr()
     assert plain_5 == plain
@@ -142,7 +142,7 @@ plain_6 = """
 A LATEXXXERROR B
 """
 
-stderr_6 = r"""*** LaTeX error: line 2, column 2:
+stderr_6 = r"""*** LaTeX error: code in 't.tex', line 2, column 2:
 *** To use cleveref with YaLafi, you need to use
 *** the 'poorman' option and \YYCleverefInput
 *** to load the sed file.
@@ -155,7 +155,7 @@ def test_6(capsys):
     capsys.readouterr()
     parms = parameters.Parameters()
     p = parser.Parser(parms, read_macros=read)
-    toks = p.parse(latex_6)
+    toks = p.parse(latex_6, source='t.tex')
     plain, pos = utils.get_txt_pos(toks)
     captured = capsys.readouterr()
     assert plain_6 == plain